backend/db.py
import mysql.connector

def get_db_connection():
    connection = mysql.connector.connect(
        host="localhost",
        user="root",
        password="pranav@/2005",
        database="event_db"
    )
    return connection
